requirements.txt
amqp==5.0.6
asgiref==3.3.4
atomicwrites==1.4.0
attrs==21.2.0
billiard==3.6.4.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Django==3.2.3
django-celery-results==2.0.1
django-debug-toolbar==3.2.1
django-tinymce==3.3.0
factory-boy==3.2.0
Faker==4.0.0
iniconfig==1.1.1
isort==5.8.0
kombu==5.1.0
mixer==7.0.6
packaging==20.9
pluggy==0.13.1
prompt-toolkit==3.0.18
py==1.10.0
pyparsing==2.4.7
pytest==6.2.4
pytest-django==4.3.0
python-dateutil==2.8.1
python-dotenv==0.17.1
pytz==2021.1
redis==3.5.3
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
